Scenarios helpers extracted
